WIP: Migration to uptown_funk macro host_functions
This is a huge change of the codebase structure, so hang in there!
This commit adds a sub-crate called uptown_funk, that exports a macro
annotation: #[host_functions(namespace = "...")]

This is the first step towards defining Wasm host functions that are
compatible with multiple runtimes (Wasmtime, Wasmer, ...). It's still
WIP and there are a few rough corners that need to be fixed before
everything works again.

Generally I'm satisfied with the outcome of it and think it's a good
foundation to start building upon.