Improve uptown_funk API.
This commit fixes many of the outstanding issues introduced by switching
to the uptown_funk crate for defining host functions.

* &mut self support for host functions.
Host functions can now take a mutable reference to their state and
directly modify it. Currently host functions are always leaf functions
and can't be called recursively, so it's safe to do so.

* Channels are stored now in a global map.
This simplifies significantly the serialisation/deserialisation of
channels.

* HashMapStore is introduced as a new structure to manage host side
state.

* WasmTo/From functions now work with u32 values instead of i32.