Reading buffer is not an Option anymore
Instead of sometimes necessary taking of the message
(because of the  mut ref to caller), we now take "data" (Vec<u8>)
from the reading message.

This can also be used to move data into draft.

The host functions have been simplified as there are less possible traps
when the reading buffer was None.

Initially there is a "message to yourself" in the reading buffer.